countries/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
djangorestframework==3.10.3
Faker==3.0.0
gunicorn==20.0.4
idna==2.8
lxml==4.4.2
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
whitenoise==4.1.4
